% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakIntoCNPs.chrom.R
\name{breakIntoCNPs.chrom}
\alias{breakIntoCNPs.chrom}
\title{TODO}
\usage{
breakIntoCNPs.chrom(segtable, chrom, startPos, endPos, startProbe,
  endProbe, eventIndex, cnptable, cnpchrom, cnpstart, cnpend, cnpindex,
  mincover, indexvals)
}
\arguments{
\item{segtable}{a \code{matrix} or a \code{data.frame} with columns 
named or enumerated by the values of 
\code{chrom, startPos, endPos, startProbe, endProbe, eventIndex}.}

\item{chrom}{a \code{character} string specifying the name for the column in 
\code{segtable} tabulating the (integer) chromosome number for each segment.}

\item{startPos}{a \code{character} string or integer specifying the 
name or number of columns in \code{segtable} that tabulates the (integer) 
genomic start coordinate of each segment.}

\item{endPos}{a \code{character} string or integer specifying the 
name or number of columns in \code{segtable} that tabulates the (integer) 
genomic end coordinate of each segment.}

\item{startProbe}{a \code{character} string specifying the names of 
columns in \code{segtable} that tabulates the (integer) start postion 
of each segment in internal units such as probe numbers for 
data of CGH microarray origin.}

\item{endProbe}{a \code{character} string specifying the names of 
columns in \code{segtable} that tabulates the (integer) end postion 
of each segment in internal units such as probe numbers for 
data of CGH microarray origin.}

\item{eventIndex}{a \code{character} string giving the name of a column in 
\code{segtable} where copy number variation status of the segments is 
tabulated.}

\item{cnptable}{a \code{matrix} or a \code{data.frame} with columns named or 
enumerated as given by \code{cnpchrom, cnpstart, cnpend, cnpindex} and 
with rows corresponding to genomic intervals that comprise the mask.}

\item{cnpchrom}{a \code{character} string or \code{integer} 
specifying the name or number of columns in \code{cnptable} that tabulates 
the chromosome number of the intervals comprising the mask.}

\item{cnpstart}{a \code{character} string or \code{integer} 
specifying the name or number of columns in \code{masktable} that tabulates 
the genomic start coordinates of the intervals comprising the mask.}

\item{cnpend}{a \code{character} string or \code{integer} 
specifying the name or number of columns in \code{masktable} that tabulates 
the genomic end coordinates of the intervals comprising the mask.}

\item{cnpindex}{a \code{numeric} \code{vector} corresponding to 
\code{eventIndex}, specifying copy number events status for measuring units.}

\item{mincover}{A single \code{numeric} value between \code{0} and \code{1} 
specifying the degree of overlap above which two clusters will be joined 
into one.}

\item{indexvals}{a \code{numeric} \code{vector} of length 2 
specifying the two values in \code{maskindex} to be matched 
with values in \code{eventIndex} to 
determine the events that are to be masked.}
}
\value{
TODO
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
