% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerprob.R
\name{centerprob}
\alias{centerprob}
\title{TODO}
\usage{
centerprob(logr, emfit, zgroup, times, center)
}
\arguments{
\item{logr}{TODO a \code{vector} of \code{numeric}}

\item{emfit}{An object of class \code{Mclust} providing a 
mixture model estimation.}

\item{times}{TODO}

\item{center}{TODO}
}
\value{
TODO a \code{vector} of \code{numeric}
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
