% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNclusterNcenter.R
\name{CNclusterNcenter}
\alias{CNclusterNcenter}
\title{TODO}
\usage{
CNclusterNcenter(
  segrat,
  blsize,
  minJoin,
  nTrial,
  bestBIC,
  modelNames,
  cweight,
  bstimes,
  chromRange
)
}
\arguments{
\item{blsize}{a single \code{integer} specifying the bootstrap 
sampling rate of segment medians to generate input for model-based 
clustering. The number of times a segment is sampled is then given by the 
(integer) division of the segment length in internal units by \code{blsize}.}

\item{minJoin}{a single \code{numeric} value between \code{0} and \code{1} 
specifying the degree of overlap above which two clusters will be joined 
into one.}

\item{nTrial}{a single \code{integer} specifying the number of times 
a model-based clustering is attempted for each profile in order to 
achieve the highest Bayesian information criterion (BIC).}

\item{bestBIC}{a single \code{numeric} value for initializing BIC 
maximization. A large negative value is recommended.}

\item{modelNames}{a \code{vector} of \code{character} strings specifying 
the names of models to be used in model-based clustering (see package 
\code{mclust} for further details).}

\item{cweight}{a single \code{numeric} value between \code{0} and \code{1} 
specifying the minimal share of the central cluster in each profile.}

\item{bstimes}{a single \code{double} value specifying the number of 
time the median of each segment is sampled in order to predict the cluster 
assignment for the segment.}

\item{chromRange}{a \code{integer} vector enumerating chromosomes from 
which segments are to be used for initial model-based clustering.}
}
\value{
a \code{matrix} with 9 columns:
\itemize{
\item{\code{untitled}}{ TODO}
\item{\code{untitled}}{ TODO}
\item{\code{mediandev}}{ a \code{numeric}, the median function of copy 
number relative to its central value}
\item{\code{segerr}}{ a \code{numeric}, the error estimate for the 
function of copy number}
\item{\code{centerz}}{ a \code{numeric} between \code{0} and \code{1}, the 
probability that the segment is in the central cluster}
\item{\code{cpb}}{ TODO}
\item{\code{maxz}}{ TODO}
\item{\code{maxzmean}}{ TODO}
\item{\code{maxzsigma}}{ TODO}
}
}
\description{
TODO
}
\examples{

# TODO

## Create a list that contain information about the segments and bins 
## related to one 



}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
