% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNprep.R
\docType{package}
\name{CNprep-package}
\alias{CNprep-package}
\alias{CNprep}
\title{CNprep: Pre-process DNA Copy Number (CN) Data for Detection of CN Events}
\description{
This package evaluates DNA copy number data, using both their initial 
form (copy number as a noisy function of genomic position) and their 
approximation by a piecewise-constant function (segmentation), for the 
purpose of identifying genomic regions where the copy number differs 
from the norm.
}
\seealso{
\itemize{
\item \code{\link{CNpreprocessing}} { for pre-processing DNA copy number 
(CN) data for detection of CN events.}
\item \code{\link{makeCNPmask}} { for creating a mask given a set of
copy number events.}
\item \code{\link{applyCNPmask}} { for applying a mask to a set of
copy number events.}
}
}
\author{
Alexander Krasnitz and Guoli Sun

Maintainer:
Guoli Sun <guolisun87@gmail.com>
}
\keyword{package}
