% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containment.indicator.R
\name{containment.indicator}
\alias{containment.indicator}
\title{TODO}
\usage{
containment.indicator(vstart, vend, wstart, wend)
}
\arguments{
\item{vstart}{TODO}

\item{vend}{TODO}

\item{wstart}{TODO}

\item{wend}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
