% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNprep.R
\docType{data}
\name{EMexample}
\alias{EMexample}
\title{A model-based clustering}
\format{an object of class \code{Mclust} providing the optimal (according 
to BIC) mixture model estimation composed of 5 clusters. 
The details of the output components are as follows:
\itemize{
\item{\code{call}}{ the matched call}
\item{\code{data}}{ a \code{matrix}, the input data matrix}
\item{\code{modelName}}{ a \code{character} string denoting the model at 
    which the optimal BIC occurs}
\item{\code{n}}{ an \code{integer}, the number of observations in the data}
\item{\code{d}}{ a \code{double}, the dimension of the data}
\item{\code{G}}{ an \code{integer}, the optimal number of mixture 
    components}
\item{\code{BIC}}{ a \code{double}, all BIC values}
\item{\code{bic}}{ a \code{double}, the optimal BIC value}
\item{\code{loglik}}{ a \code{double}, the log-likelihood corresponding 
    to the optimal BIC}
\item{\code{df}}{ a \code{double}, the number of estimated parameters}
\item{\code{hypvol}}{ \code{NA}}
\item{\code{parameters}}{ a \code{list} with the following components:
\itemize{
    \item{\code{pro}}{ a \code{vector} of \code{double} whose \code{k}th 
    component is the mixing proportion for the \code{k}th component 
    of the mixture model}
    \item{\code{mean}}{ a \code{matrix} of \code{double}  whose \code{k}th 
    column is the mean of the \code{k}th component of the mixture model}
    \item{\code{variance}}{ a \code{list} of variance parameters for 
    the model}
}
}
\item{\code{z}}{ a \code{matrix} of \code{double}  whose 
    \code{[i,k]}th entry is the probability that observation \code{i} in 
    the test data belongs to the \code{k}th class}
\item{\code{classification}}{ an \code{array} of \code{double}, the 
    classification corresponding to \code{z}}
\item{\code{uncertainty}}{ a \code{double}, the uncertainty associated 
    with the classification}
}}
\usage{
data(EMexample)
}
\value{
an object of class \code{Mclust} providing the optimal (according 
to BIC) mixture model estimation composed of 5 clusters. 
The details of the output components are as follows:
\itemize{
\item{\code{call}}{ the matched call}
\item{\code{data}}{ a \code{matrix}, the input data matrix}
\item{\code{modelName}}{ a \code{character} string denoting the model at 
    which the optimal BIC occurs}
\item{\code{n}}{ an \code{integer}, the number of observations in the data}
\item{\code{d}}{ a \code{double}, the dimension of the data}
\item{\code{G}}{ an \code{integer}, the optimal number of mixture 
    components}
\item{\code{BIC}}{ a \code{double}, all BIC values}
\item{\code{bic}}{ a \code{double}, the optimal BIC value}
\item{\code{loglik}}{ a \code{double}, the log-likelihood corresponding 
    to the optimal BIC}
\item{\code{df}}{ a \code{double}, the number of estimated parameters}
\item{\code{hypvol}}{ \code{NA}}
\item{\code{parameters}}{ a \code{list} with the following components:
\itemize{
    \item{\code{pro}}{ a \code{vector} of \code{double} whose \code{k}th 
    component is the mixing proportion for the \code{k}th component 
    of the mixture model}
    \item{\code{mean}}{ a \code{matrix} of \code{double}  whose \code{k}th 
    column is the mean of the \code{k}th component of the mixture model}
    \item{\code{variance}}{ a \code{list} of variance parameters for 
    the model}
}
}
\item{\code{z}}{ a \code{matrix} of \code{double}  whose 
    \code{[i,k]}th entry is the probability that observation \code{i} in 
    the test data belongs to the \code{k}th class}
\item{\code{classification}}{ an \code{array} of \code{double}, the 
    classification corresponding to \code{z}}
\item{\code{uncertainty}}{ a \code{double}, the uncertainty associated 
    with the classification}
}
}
\description{
A model-based clustering based on parameterized finite Gaussian 
mixture models
}
\details{
See \code{\link[mclust]{Mclust}} for detailed description of
the object of class \code{mclust}.
}
\examples{

## Loading the demo object of class 'Mclust'
data(EMexample)

## Group clusters that have at least 2\% of overlap
## The inital object has 5 clusters while the return object has only 
## 4 clusters
CNprep:::consolidate(EMexample, minover=0.2)


}
\seealso{
\itemize{
\item \code{\link{CNpreprocessing}} {for pre-process DNA copy number (CN) 
data for detection of CN events.}
}
}
\keyword{datasets}
