% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCNPmask.chrom.R
\name{makeCNPmask.chrom}
\alias{makeCNPmask.chrom}
\title{Given a set of copy-number events, create a DNA copy-number mask for
one specific chromosome.}
\usage{
makeCNPmask.chrom(imat, startcol = 1, endcol = 2, nprof = 1, uthresh,
  dthresh)
}
\arguments{
\item{imat}{A \code{matrix} or a \code{data.frame} tabulating the chromosome 
numbers and endpoint positions of the interval events.}

\item{startcol}{A \code{character} string or \code{numeric} (representing an
integer) specifying the column of \code{imat} containing the left 
endpoint of the interval events. Default: \code{1}.}

\item{endcol}{A \code{character} string or \code{numeric} (representing an
integer) specifying the column of \code{imat} containing the right endpoint 
of the interval events. Default: \code{2}.}

\item{nprof}{A \code{numeric} acting as an integer specifying the number 
of copy number profiles from which the events originated. Default: \code{1}.}

\item{uthresh}{A \code{numeric} specifying the upper threshold for 
the event frequency or (if \code{nprof = 1}) for the event count.}

\item{dthresh}{A \code{numeric} specifying the upper and lower thresholds for 
the event frequency or (if \code{nprof = 1}) for the event count.}
}
\value{
TODO
}
\description{
The function takes as an input a set of intervals with 
integer-valued boundary positions. It then finds interval regions where 
the event count is above each of two thresholds, upper and lower, and 
returns those interval regions with the count above the lower 
threshold that contain interval regions with the count above the upper 
threshold. This is done for one specific chromosome.
}
\examples{

# TODO

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
