% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalComparison.R
\name{normalComparison}
\alias{normalComparison}
\title{TODO}
\usage{
normalComparison(normalmedian, normalength, tumormedian, tumorlength,
  normalmad = NULL, normalerror = NULL, tumormad = NULL,
  tumorerror = NULL)
}
\arguments{
\item{normalmedian}{a \code{matrix} of \code{double} of one column. TODO}

\item{normalength}{a \code{matrix} of \code{double} of one column. TODO}

\item{tumormedian}{a \code{vector} of \code{double}. TODO}

\item{tumorlength}{a \code{vector} of \code{double}. Default: \code{NULL}.}

\item{normalmad}{TODO. Default: \code{NULL}.}

\item{normalerror}{TODO. Default: \code{NULL}.}

\item{tumormad}{a \code{vector} of \code{double}. Default: \code{NULL}.}

\item{tumorerror}{TODO.  Default: \code{NULL}.}
}
\value{
TODO
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
