% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segsample.R
\name{segsample}
\alias{segsample}
\title{TODO}
\usage{
segsample(mysegs, ratcol, startcol = "StartProbe", endcol = "EndProbe",
  blocksize = 0, times = 0)
}
\arguments{
\item{mysegs}{TODO}

\item{ratcol}{TODO}

\item{startcol}{TODO}

\item{endcol}{TODO}

\item{blocksize}{TODO}

\item{times}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
