% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smad.R
\name{smad}
\alias{smad}
\title{TODO}
\usage{
smad(pos, v)
}
\arguments{
\item{pos}{TODO}

\item{v}{TODO}
}
\value{
TODO
}
\description{
Compute the median absolute deviation TODO
}
\examples{

# TODO

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
