% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smedian.R
\name{smedian}
\alias{smedian}
\title{TODO}
\usage{
smedian(pos, v)
}
\arguments{
\item{pos}{TODO}

\item{v}{TODO}
}
\value{
TODO
}
\description{
Compute the median TODO
}
\examples{

# TODO

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
