% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smedian.sample.R
\name{smedian.sample}
\alias{smedian.sample}
\title{TODO}
\usage{
smedian.sample(pos, v)
}
\arguments{
\item{pos}{TODO}

\item{v}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

position <- c(1, 10)
names(position) <- c("StartProbe", "EndProbe")

values <- c(0.072073840, 0.119913919, 0.154459489, 0.040994620, -0.082843732,
    0.093052725, 0.170908930, 0.086624752, -0.003855011, -0.195791649,
    0.012126180, 0.043428961, 0.028435453, 0.075708220, 0.020358061)

CNprep:::smedian.sample(pos=position, v=values)

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
