% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smedmad.R
\name{smedmad}
\alias{smedmad}
\title{TODO}
\usage{
smedmad(pos, v)
}
\arguments{
\item{pos}{TODO}

\item{v}{TODO}
}
\value{
TODO
}
\description{
Compute the median and the median absolute deviation,TODO
}
\examples{

# TODO

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
