% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateMakeCNPmask.R
\name{validateMakeCNPmask}
\alias{validateMakeCNPmask}
\title{Parameters validation for the \code{\link{makeCNPmask}} function}
\usage{
validateMakeCNPmask(imat, chromCol, startCol, endCol, nProf, uThresh,
  dThresh)
}
\arguments{
\item{imat}{A \code{matrix} or a \code{data.frame} tabulating the chromosome 
numbers and startpoint/endpoint positions of the interval events.}

\item{chromCol}{A \code{character} string or \code{numeric} representing an
integer specifying the column of \code{imat} containing the chromosome 
numbers of the interval events.}

\item{startCol}{A \code{character} string or \code{numeric} representing an
integer   
specifying the column of \code{imat} containing the left (right) endpoint of
the interval events.}

\item{endCol}{A \code{character} string or \code{numeric} representing an
integer specifying the column of \code{imat} containing the right endpoint 
of the interval events.}

\item{nProf}{A positive \code{numeric} acting as an integer specifying 
the number of copy number profiles from which the events originated.}

\item{uThresh}{A \code{numeric} specifying the upper ratio threshold for 
the event frequency or (if \code{nProf = 1}) for the event count.
The upper ratio threshold must be equal or superior to the lower 
threshold (\code{dThresh}). The upper ratio threshold must also be between
0 and 1.}

\item{dThresh}{A \code{numeric} specifying the lower ratio threshold 
for the event frequency or (if \code{nProf = 1}) for the event count.
The lower ratio threshold must be equal or inferior to the upper  
threshold (\code{dThresh}). 
The lower ratio threshold must also be between 0 and 1.}
}
\value{
\code{0}.
}
\description{
Validation of all parameters needed by the public
\code{\link{makeCNPmask}} function.
}
\examples{

data(segexample)

## Return zero as all parameters are valid
CNprep:::makeCNPmask(imat=segexample, chromCol="chrom",
    startCol="start", endCol="end", nProf=1203, uThresh=0.20, dThresh=0.10)

}
\author{
Astrid Deschênes
}
\keyword{internal}
