% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.median.R
\name{weighted.median}
\alias{weighted.median}
\title{Calculate the lower weighted median of a specified vector.}
\usage{
weighted.median(v, weights)
}
\arguments{
\item{v}{a \code{vector} of \code{double} containing the values used for
the calculation.}

\item{weights}{a \code{vector} of \code{double} containing the weight for 
each value of the \code{v} vector.}
}
\value{
a \code{double} which is the weighted median of the vector.
}
\description{
Compute the lower weighted median of a specified vector.
}
\examples{

## Vector of values
values <- c(0.06363411, 0.04342896, 0.07207384, 0.07319237, 0.07273546,
    0.01463932, 0.02136043, 0.01967027)

## Vector of weigth associated to each value    
weight <- c(1, 1, 1, 1, 0, 0, 0, 0)

## Calculation of the weighted median
CNprep:::weighted.median(v = values, weights = weight)

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
