% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNprep.R
\docType{data}
\name{annotexample}
\alias{annotexample}
\title{Annotation table for ROMA CGH platform and human genome 
version 17.}
\format{A data frame with 83055 observations on the following 3 variables.
\describe{
\item{\code{PROBEID}}{a character vector}
\item{\code{CHROM}}{a numeric vector}
\item{\code{CHROM.POS}}{a numeric vector}
}}
\source{
GEO accession GPL9775, 
http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GPL9775

GEO accession GPL9775, 
http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GPL9775
}
\usage{
data(annotexample)

data(annotexample)
}
\value{
A data frame with 83055 observations on the following 3 variables.
\describe{
\item{\code{PROBEID}}{a character vector}
\item{\code{CHROM}}{a numeric vector}
\item{\code{CHROM.POS}}{a numeric vector}
}

a \code{data.frame} with 83055 observations on the following 3 
variables. 
\describe{
\item{\code{PROBEID}}{a \code{character} \code{vector} of probe names.}
\item{\code{CHROM}}{a \code{numeric} \code{vector} of chromosome positions.}
\item{\code{CHROM.POS}}{a \code{numeric} \code{vector} of genomic 
positions.}
}
}
\description{
Whole genome annotation table using Representational 
Oligonucleotide Microarray Analysis (ROMA) CGH platform, 
human genome version 17.

Whole genome annotation table using Representational 
Oligonucleotide Microarray Analysis (ROMA) CGH platform, 
human genome version 17.
}
\details{
The values in the chromosome column are all integer, with 
23 corresponding to X, 24 to Y and 25 to a set of non-human test probes.

The values in the chromosome column are all integer, with 
23 corresponding to X, 24 to Y and 25 to a set of non-human test probes.
}
\examples{

## Loading annotation table dataset
data(annotexample)

## Loading other datasets
data(ratexample)
data(segexample)
data(normsegs)

## How to use annotexample, when segment table does not have columns 
## of integer postions in terms of measuring units(probes), such 
## as "mysegs" below 

mysegs<-segexample[,c(1,5:12)]

## Analysis limited to chromosomes 1 and 2
## The bstimes variable should be higher for a real analysis
##segtable <- CNpreprocessing(segall=mysegs, ratall=ratexample, "ID",
##    chromCol="chrom", bpstartCol="chrom.pos.start", bpendCol="chrom.pos.end",
##    annot=annotexample, annotStartCol="CHROM.POS", annotEndCol="CHROM.POS",
##   annotChromCol="CHROM", blsize=50, minjoin=0.25, cWeight=0.4, bstimes=5,
##    chromrange=1:2, nJobs=1, modelNames="E", 
##    normalLength=normsegs[,1], normalMedian=normsegs[,2])


## Loading annotation table dataset
data(annotexample)

## Loading other datasets
data(ratexample)
data(segexample)
data(normsegs)

## How to use annotexample, when segment table does not have columns 
## of integer postions in terms of measuring units(probes), such 
## as "mysegs" below 

mysegs <- segexample[,c(1,5:12)]

## Analysis limited to chromosomes 1 and 2
## The bstimes variable should be higher for a real analysis
segtable <- CNpreprocessing(segall=mysegs, ratall=ratexample, idCol="ID",
    chromCol="chrom", bpstartCol="chrom.pos.start", bpendCol="chrom.pos.end",
    annot=annotexample, annotStartCol="CHROM.POS", annotEndCol="CHROM.POS",
    annotChromCol="CHROM", blsize=50, minJoin=0.25, cWeight=0.4, bstimes=3,
    chromRange=1:2, nJobs=1, modelNames="E", 
    normalLength=normsegs[,1], normalMedian=normsegs[,2])

}
\seealso{
\itemize{
\item \code{\link{CNpreprocessing}} {for pre-process DNA copy number (CN) 
data for detection of CN events.}
}

\itemize{
\item \code{\link{CNpreprocessing}} {for pre-process DNA copy number (CN) 
data for detection of CN events.}
}
}
\keyword{datasets}
