% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakIntoGaps.R
\name{breakIntoGaps}
\alias{breakIntoGaps}
\title{TODO}
\usage{
breakIntoGaps(segtable, gapind, StartProbe, EndProbe)
}
\arguments{
\item{segtable}{TODO}

\item{gapind}{TODO}

\item{StartProbe}{TODO}

\item{EndProbe}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
