% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakIntoGaps.R
\name{breakIntoGaps}
\alias{breakIntoGaps}
\title{TODO}
\usage{
breakIntoGaps(segtable, gapind, StartProbe, EndProbe)
}
\arguments{
\item{segtable}{a \code{matrix} or a \code{data.frame} TODO}

\item{gapind}{a \code{character} string TODO}

\item{StartProbe}{a \code{character} string specifying the names of 
columns in \code{segtable} that tabulates the (integer) start postion 
of each segment in internal units such as probe numbers for 
data of CGH microarray origin.}

\item{EndProbe}{a \code{character} string specifying the names of 
columns in \code{segtable} that tabulates the (integer) end postion 
of each segment in internal units such as probe numbers for 
data of CGH microarray origin.}
}
\value{
TODO
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
