% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerprob.R
\name{centerprob}
\alias{centerprob}
\title{TODO}
\usage{
centerprob(logr, emfit, zgroup, times, center)
}
\arguments{
\item{logr}{TODO a \code{vector} of \code{numeric}}

\item{emfit}{an object of class \code{Mclust} providing a 
mixture model estimation.}

\item{times}{a single \code{numeric} value, used as an integer, specifying 
the number of time the median of each segment is sampled in order to 
predict the cluster assignment for the segment.}

\item{center}{TODO}
}
\value{
TODO a \code{vector} of \code{numeric}
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
