% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerprob.R
\name{centerprob}
\alias{centerprob}
\title{TODO}
\usage{
centerprob(logr, emfit, zgroup, times, center)
}
\arguments{
\item{pos}{TODO}

\item{v}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
