% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerprob.R
\name{centerprob}
\alias{centerprob}
\title{TODO}
\usage{
centerprob(logr, emfit, zgroup, times, center)
}
\arguments{
\item{logr}{a \code{vector} of \code{double} TODO}

\item{emfit}{an object of class \code{Mclust} providing a 
mixture model estimation for the clusters.}

\item{times}{a \code{integer} TODO}

\item{center}{a \code{integer} TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
