% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNprep.R
\docType{data}
\name{cnpexample}
\alias{cnpexample}
\title{Example of a boundary positions table.}
\format{A data frame with 19188 rows and 4 columns. 
\describe{
\item{\code{copy.num}}{a character vector indicating whether an event is
a gain ("amp") or a loss ("del"). }
\item{\code{chrom}}{a numeric vector indicating which chromosome the 
event is in.}
\item{\code{chrom.start}}{a numeric vector of event start positions.}
\item{\code{chrom.end}}{a numeric vector of event start positions.}
}}
\source{
Strong association of de novo copy number mutations with autism.
Sebat J, Lakshmi B, Malhotra D, Troge J, Lese-Martin C, Walsh T, Yamrom B, 
Yoon S, Krasnitz A, Kendall J, Leotta A, Pai D, Zhang R, Lee YH, Hicks J, 
Bregman J, Sutcliffe JS, Jobanputra V, Chung W, Warburton D, King MC, 
Skuse D, Geschwind DH, Gilliam TC, Ye K, Wigler M. 
Science. 2007 Apr 20;316(5823):445-9.
}
\usage{
data(cnpexample)
}
\value{
A data frame with 19188 rows and 4 columns. 
\describe{
\item{\code{copy.num}}{a character vector indicating whether an event is
a gain ("amp") or a loss ("del"). }
\item{\code{chrom}}{a numeric vector indicating which chromosome the 
event is in.}
\item{\code{chrom.start}}{a numeric vector of event start positions.}
\item{\code{chrom.end}}{a numeric vector of event start positions.}
}
}
\description{
A table of genomic positions for DNA copy-number changing 
events, collected from genomes of 1203 individuals using Representational 
Oligonucleotide Microarray Analysis (ROMA) platform.
}
\examples{

## Loading dataset
data(cnpexample)

## Create masked table usign cnpindex vector 
##namps17 <- cnpexample[cnpexample[,"copy.num"]=="amp",] 
##aCNPmask <- makeCNPmask(imat=namps17, chromCol=2, startCol=3, endCol=4, 
##    nProf=1203, uthresh=0.02, dthresh=0.008) 
##ndels17 <- cnpexample[cnpexample[,"copy.num"]=="del",] 
##dCNPmask <- makeCNPmask(imat=ndels17, chromCol=2, startCol=3, endCol=4, 
##    nProf=1203, uthresh=0.02, dthresh=0.008) 
##cnptable <- rbind(cbind(aCNPmask,cnpindex=1),cbind(dCNPmask,cnpindex=-1))

# TODO
## Run the CNP test using masked table
#myCNPtable <- applyCNPmask(segtable,"chrom",startPos="chrom.pos.start", 
#    endPos="chrom.pos.end","start","end","eventIndex",masktable=cnptable,
#    "chrom", maskstart="start",maskend="end",maskindex="cnpindex",
#    mincover=0.005,indexvals=c(-1,1))

}
\seealso{
\itemize{
\item \code{\link{applyCNPmask}} {for applying a mask to a table of 
copy number events.}
}
}
\keyword{datasets}
