% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.center.R
\name{get.center}
\alias{get.center}
\title{TODO}
\usage{
get.center(emfit, mincenter)
}
\arguments{
\item{emfit}{An object of class \code{Mclust} providing a 
mixture model estimation.}

\item{mincenter}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
