% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.center.R
\name{get.center}
\alias{get.center}
\title{TODO}
\usage{
get.center(emfit, mincenter)
}
\arguments{
\item{emfit}{a \code{list} containing information about the updated mixture
model for the clusters as given by an object of class 
\code{Mclust}.}

\item{mincenter}{TODO}
}
\description{
TODO
}
\details{
The \code{emfit} parameter correspond to an object of class 
\code{Mclust} that has been modified by the \code{CNprep:::consolidate}
function.
}
\examples{

# TODO

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
