% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getz.R
\name{getz}
\alias{getz}
\title{TODO}
\usage{
getz(logr, emfit, zgroup, times)
}
\arguments{
\item{logr}{TODO}

\item{emfit}{An object of class \code{Mclust} providing a 
mixture model estimation.}

\item{zgroup}{TODO}

\item{times}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
