% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCNPmask.chrom.R
\name{makeCNPmask.chrom}
\alias{makeCNPmask.chrom}
\title{TODO}
\usage{
makeCNPmask.chrom(imat, startcol = 1, endcol = 2, nprof = 1, uthresh,
  dthresh)
}
\arguments{
\item{imat}{TODO}

\item{startcol}{TODO}

\item{endcol}{TODO}

\item{nprof}{TODO}

\item{uthresh}{TODO}

\item{dthresh}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
