% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalComparison.R
\name{normalComparison}
\alias{normalComparison}
\title{TODO}
\usage{
normalComparison(normalmedian, normalength, tumormedian, tumorlength,
  normalmad = NULL, normalerror = NULL, tumormad = NULL,
  tumorerror = NULL)
}
\arguments{
\item{normalmedian}{TODO}

\item{normalength}{TODO}

\item{tumormedian}{TODO}

\item{tumorlength}{TODO. Default: \code{NULL}.}

\item{normalmad}{TODO. Default: \code{NULL}.}

\item{normalerror}{TODO. Default: \code{NULL}.}

\item{tumormad}{TODO. Default: \code{NULL}.}

\item{tumorerror}{TODO. Default: \code{NULL}.}
}
\value{
TODO
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
