% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalComparison.R
\name{normalComparison}
\alias{normalComparison}
\title{TODO}
\usage{
normalComparison(
  normalmedian,
  normalength,
  tumormedian,
  tumorlength,
  normalmad = NULL,
  normalerror = NULL,
  tumormad = NULL,
  tumorerror = NULL
)
}
\arguments{
\item{normalmedian}{a \code{numeric} \code{vector}, 
of the same length as \code{normalength}, specifying the segment values
of the normal reference segments.}

\item{normalength}{an integer \code{vector} specifying the genomic lengths 
of segments in the normal reference data.}

\item{tumormedian}{TODO}

\item{tumorlength}{TODO. Default: \code{NULL}.}

\item{normalmad}{a \code{numeric} \code{vector}, 
of the same length as \code{normalength}, specifying the value spreads 
of the normal reference segments. Default: \code{NULL}.}

\item{normalerror}{a numeric \code{vector}, 
of the same length as \code{normalength}, specifying the error values
of the normal reference segments.Default: \code{NULL}.}

\item{tumormad}{TODO. Default: \code{NULL}.}

\item{tumorerror}{TODO. Default: \code{NULL}.}
}
\value{
TODO
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
