% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNprep.R
\docType{data}
\name{normsegs}
\alias{normsegs}
\title{A reference set of segments}
\format{
a \code{matrix} with 43497 rows/segments and 2 columns/variables. 
The 2 columns are:
\itemize{
\item{\code{length}}{ a \code{numeric} \code{vector}, used as integer, of 
    segment genomic length}
\item{\code{segmedian}}{ a \code{numeric} \code{vector} of segment 
    median computed from log copy number ratio}
}
}
\source{
Sebat J, et al. Strong association of de novo copy number 
mutations with autism. Science. 2007 Apr 20;316(5823):445-9. 
Epub 2007 Mar 15.
}
\usage{
data(normsegs)
}
\value{
a \code{matrix} with 43497 rows/segments and 2 columns/variables. 
The 2 columns are:
\itemize{
\item{\code{length}}{ a \code{numeric} \code{vector}, used as integer, of 
    segment genomic length}
\item{\code{segmedian}}{ a \code{numeric} \code{vector} of segment 
    median computed from log copy number ratio}
}
}
\description{
A table of segment lengths and log copy number ratios for 
a large set of human diploid genomes.
}
\details{
The table originates in a set of copy number profiles of over 
1000 individuals, obtained using Representational Oligonucleotide Microarray 
Analysis (ROMA) technology. To ensure ploidy of 2 segments from X and Y 
chromosomes and segments shorter than 5Mb were excluded.
}
\examples{

## Loading log ratio dataset
data(segexample)

## Load datasets
data(segexample)
data(ratexample)
data(normsegs)

## Preprocess segments for WZ3 sample
segtable <- CNpreprocessing(segall=segexample[segexample[,"ID"] == "WZ3",],
                 ratall=ratexample, idCol="ID", startCol = "start", 
                 endCol="end", chromCol="chrom", 
                 bpStartCol="chrom.pos.start", 
                 bpEndCol="chrom.pos.end", blsize=50, 
                 minJoin=0.30, cWeight=0.4,
                 bsTimes = 40, chromRange=1:22, nJobs=1, 
                 modelNames="E", normalLength=normsegs[,1], 
                 normalMedian=normsegs[,2])

}
\seealso{
\itemize{
\item \code{\link{CNpreprocessing}} {for pre-process DNA copy number (CN) 
data for detection of CN events.}
}
}
\keyword{datasets}
