% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNprep.R
\docType{data}
\name{ratexample}
\alias{ratexample}
\title{Example of copy number log ratio dataset}
\format{a log ratio \code{matrix} with 83055 rows, one per 
oligonucleotide probe, and 5 columns, one for each breast tumor sample.}
\source{
Hicks, J. et al. Novel patterns of genome rearrangement and their 
association with survival in breast cancer. Genome Res. 2006. 16:1465–1479.
doi: 10.1101/gr.5460106
}
\usage{
data(ratexample)
}
\value{
a log ratio \code{matrix} with 83055 rows, one per 
oligonucleotide probe, and 5 columns, one for each breast tumor sample.
}
\description{
Log ratio data for 5 breast cancer genomes, derived using 
Representational Oligonucleotide Microarray Analysis (ROMA), an array-based 
hybridization method that uses genomic complexity reduction based 
on representations.
}
\details{
The values are natural log copy number ratios, consistent with 
data in \code{segexample} (segmented data for these tumors) and 
\code{normsegs}. These copy number ratios are normalized using an 
intensity-based lowess curve fitting algorithm.
}
\examples{

## Loading log ratio dataset
data(ratexample)

## Plot the whole genome log ratio data for the first profile "WZ1"
## Note X and Y chromosomes at the far right of the plot
plot(ratexample[,"WZ1"], ylab="log ratio", xlab="position")

}
\seealso{
\itemize{
\item \code{\link{CNpreprocessing}} {for pre-process DNA copy number (CN) 
data for detection of CN events.}
}
}
\keyword{datasets}
