% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segsample.R
\name{segsample}
\alias{segsample}
\title{TODO}
\usage{
segsample(mysegs, ratcol, startcol = "StartProbe", endcol = "EndProbe",
  blocksize = 0, times = 0)
}
\arguments{
\item{mysegs}{a \code{data.frame} TODO}

\item{ratcol}{TODO}

\item{startcol}{a \code{character} string specifying the name of column 
in \code{mysegs} that tabulates the (integer) start postion of each segment 
in internal units such as probe numbers for data of CGH microarray origin.
Default: "StartProbe".}

\item{endcol}{a \code{character} string specifying the name of column 
in \code{mysegs} that tabulates the (integer) end postion of each segment 
in internal units such as probe numbers for data of CGH microarray origin.
Default: "EndProbe".}

\item{blocksize}{TODO . Default: \code{0}.}

\item{times}{TODO . Default: \code{0}.}
}
\value{
TODO
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
