% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smedian.sample.R
\name{smedian.sample}
\alias{smedian.sample}
\title{TODO}
\usage{
smedian.sample(pos, v)
}
\arguments{
\item{pos}{TODO}

\item{v}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Alexander Krasnitz, Guoli Sun
}
\keyword{internal}
